Stop Sync Start

+++ src/reboot.rs
@@ -1,9 +1,12 @@
 use crate::term;
 
-pub fn main(error_folder: &String) {
+pub fn main(error_folder: &String, dry_run: bool) {
     println!("reboot: working...");
 
-    term::exec(error_folder, "reboot", vec![], "reboot server");
+    if dry_run {
+    } else {
+        term::exec(error_folder, "reboot", vec![], "reboot server");
+    }
 
     println!("reboot: done! (might take some time until the reboot it performed)");
 }
